Adds ability to cache responses from PhantomJS
It's not super practical to use PhantomJS to render every page. It's performance
implications are just to great to do in production. This commit helps the issue
somewhat by adding a cahcing layer in front of phantom rendering. The cache is
keyed by the post-transform URL.

To enable caching, pass in a config object compatible with
[cache-client](https://www.npmjs.com/package/cache-client).